fix ruff and document
